refactor(web): extract pure helpers from page.tsx
Move pending-message sessionStorage helpers into lib/pending-message.ts and
the sidebar chat-tree ordering + next-chat-after-deletion logic into
lib/chat-tree.ts. Both are now pure, dependency-light, and unit-testable.
